Move updater-notifier attachment to manager
Prior to this, the updater attached itself to its notifiers. This was
not the best design since it required the updater to handle config that
was not its own (notifier names and globally configured notifiers), and
it also made mocking an updater for tests harder.

This moves the updater-notifier attachment to the manager, which already
has the global configuration data.

It also contains some bug fixes for module names.